Update the link to HERE's terms
The link has changed, so update to the new link.

Also make the link open in a new window as it's just to provide further
information.  Note the use of `rel="noopener noreferrer"` for security
as described on [MDN](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a#attr-target).